permissions work on read and update on backend

+++ pages/api/documents/[id].ts
@@ -1,58 +1,66 @@
 import { getSession } from '@auth0/nextjs-auth0'
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { deleteDocument, deletePermission, deletePermissionByDoc, getDocument, updateDocument } from "../../../lib/mongoUtils"
-import { DocumentData } from '../../../types/globals'
+import { deleteDocument, deletePermissionByDoc, getDocument, getPermissionByDoc, updateDocument } from "../../../lib/mongoUtils"
+import { DocumentData, PermissionData, UserPermission } from '../../../types/globals'
 
 export default async function documentHandler(req: NextApiRequest, res: NextApiResponse) {
   const { query, method } = req
   const session = getSession(req, res)
 
+  const permissions = await getPermissionByDoc(query.id.toString()) as PermissionData
+  const isOwner = permissions.ownerId === session?.user.sub
+  const isRestricted = permissions.globalPermission === UserPermission.None
+  const user = permissions.users.find((user) => user.email === session?.user.email)
+
+  const canComment = user && [UserPermission.Comment, UserPermission.Edit].indexOf(user.permission) > -1 || isOwner
+  const canEdit = user && user.permission === UserPermission.Edit || isOwner
+  const globalEdit = [UserPermission.Comment, UserPermission.Edit].indexOf(permissions.globalPermission) > -1
+  const globalComment = permissions.globalPermission === UserPermission.Edit
+
   switch (method) {
     case 'GET':
       const document = await getDocument(query.id.toString()) as DocumentData
-      const isRestricted = document.view.length > 0
-      const inView = document.view.includes(session?.user.email)
-      const inComment = document.comment.includes(session?.user.email)
-      const inEdit = document.edit.includes(session?.user.email)
-      const userHasPermission = inView || inComment || inEdit
 
       if (isRestricted) {
-        if (userHasPermission) {
-          document.canComment = inEdit || inComment
-          document.canEdit = inEdit
-
+        if (user || isOwner) {
+          document.canComment = canComment
+          document.canEdit = canEdit
           return res.status(200).json(document)
-        } else {
+        } 
+        else {
           return res.status(400).send({ error: 'you do not have the permissions to view this file' })
         }
-      }
-      document.canComment = document.comment.length === 0
-      document.canEdit = document.edit.length === 0
+      } 
 
+      document.canComment = globalComment
+      document.canEdit = globalEdit
       res.status(200).json(document)
       break
     case 'PATCH':
-      const oGDoc = await getDocument(query.id.toString()) as DocumentData
-      const isOGRestricted = oGDoc.view.length > 0
-      const inOGComment = oGDoc.comment.includes(session?.user.email)
-      const inOGEdit = oGDoc.edit.includes(session?.user.email)
-      
-      if (isOGRestricted) {
-        if (inOGComment || inOGEdit) {
+      if (isRestricted) {
+        if (canComment || canEdit || isOwner) {
           const updatedDocument = await updateDocument(query.id.toString(), req.body) as DocumentData
           res.status(200).json(updatedDocument)
         } else {
           return res.status(400).send({ error: 'you do not have the permissions to modify this file' })
         }
+      } 
+      if (globalEdit || globalComment) {
+        const updatedDocument = await updateDocument(query.id.toString(), req.body) as DocumentData
+        return res.status(200).json(updatedDocument)
+      } else {
+        return res.status(400).send({ error: 'you do not have the permissions to modify this file' })
       }
-      const updatedDocument = await updateDocument(query.id.toString(), req.body) as DocumentData
-      res.status(200).json(updatedDocument)
       break
     case 'DELETE':
-      await deleteDocument(query.id.toString())
-      await deletePermissionByDoc(query.id.toString())
-
-      res.status(200).json('document deleted')
+      if (isOwner) {
+        await deleteDocument(query.id.toString())
+        await deletePermissionByDoc(query.id.toString())
+        res.status(200).json('document deleted')
+      } else {
+        return res.status(400).send({ error: 'you do not have the permissions to delete this file' })
+      }
+      
       break
     default:
       res.setHeader('Allow', ['GET', 'PATCH', 'DELETE'])
